Revert "Remove stop functionality"
This reverts commit 248ec29ed436a2c6f6678a63cbc7a06ea02a1641.

+++ hyrisecockpit/database_manager/worker_pool.py
@@ -70,21 +70,26 @@ class WorkerPool:
         )
 
     def _init_workers(self) -> None:
-        self._execute_task_worker_done_event = (
-            self._generate_execute_task_worker_done_events()
-        )
-        self._execute_task_workers = self._generate_execute_task_worker()
-        self._fill_task_worker = self._generate_fill_task_worker()
+        if len(self._execute_task_workers) == 0:
+            self._execute_task_worker_done_event = (
+                self._generate_execute_task_worker_done_events()
+            )
+            self._execute_task_workers = self._generate_execute_task_worker()
+        if not self._fill_task_worker:
+            self._fill_task_worker = self._generate_fill_task_worker()
 
     def _terminate_worker(self) -> None:
-        self._fill_task_worker.terminate()  # type: ignore
-        self._fill_task_worker = None
-        for i in range(self._number_worker):
-            self._execute_task_workers[i].terminate()
-        self._execute_task_workers = []
-        self._worker_continue_event = Event()
-        self._task_queue = Queue(0)
-        self._failed_task_queue = Queue(0)
+        if not self._status == "closed":
+            if self._fill_task_worker:
+                self._fill_task_worker.terminate()
+            self._fill_task_worker = None
+            for i in range(self._number_worker):
+                self._execute_task_workers[i].terminate()
+            self._execute_task_workers = []
+            self._worker_continue_event = Event()
+            self._task_queue = Queue(0)
+            self._failed_task_queue = Queue(0)
+            self._status = "closed"
 
     def _wait_for_worker(self) -> None:
         self._worker_continue_event.clear()
@@ -94,7 +99,8 @@ class WorkerPool:
 
     def _start_worker(self) -> None:
         self._continue_execution_flag.value = True
-        self._fill_task_worker.start()  # type: ignore
+        if self._fill_task_worker:
+            self._fill_task_worker.start()
         for i in range(self._number_worker):
             self._execute_task_workers[i].start()
 
@@ -104,13 +110,25 @@ class WorkerPool:
             self._init_workers()
             self._start_worker()
             self._status = "running"
+        elif self._status == "stopped":
+            self._continue_execution_flag.value = True
+            self._worker_continue_event.set()
+            self._status = "running"
+        self._database_blocked.value = False
+
+    def _stop_job(self) -> None:
+        if self._status == "running":
+            self._wait_for_worker()
+            self._status = "stopped"
         self._database_blocked.value = False
 
     def _close_job(self) -> None:
-        if self._status == "running":
+        if self._status == "stopped":
+            self._terminate_worker()
+        elif self._status == "running":
             self._wait_for_worker()
             self._terminate_worker()
-        self._status = "closed"
+        self._status == "closed"
         self._database_blocked.value = False
 
     def start(self) -> bool:
@@ -118,6 +136,15 @@ class WorkerPool:
         if not self._database_blocked.value:
             self._database_blocked.value = True
             self._scheduler.add_job(func=self._start_job)
+            return True
+        else:
+            return False
+
+    def stop(self) -> bool:
+        """Stop worker."""
+        if not self._database_blocked.value:
+            self._database_blocked.value = True
+            self._scheduler.add_job(func=self._stop_job)
             return True
         else:
             return False
